+++ HISTORY.md
@@ -8,6 +8,7 @@ unreleased
     - deps: parseurl@~1.3.3
     - deps: statuses@~1.4.0
   * deps: parseurl@~1.3.3
+  * perf: remove substr call from FQDN mapping
 
 3.6.6 / 2018-02-14
 ==================